backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Fast API Development"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "fastapi>=0.116.1",
    "isort>=6.0.1",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
    "pyjwt>=2.10.1",
    "passlib[bcrypt]>=1.7.4",
]
